feat: Added experiments with modification of the baseline
In this case, the validation set is kept constant for every experiment.
The images in the validation set have been modified previously, therefore,
no data augmentation is performed on them when training the model.
The augmentation in the validation was hard, meaning that the transformations
were done over another range of values. For instance, the rotation was
performed between 30 and 50. In the other hand, the data augmentation
performed in the training set is soft, for instance, the rotation is
done between 15 and 35.

The idea of the same validation set is to make the models more
comparable since they are being validated against the very same
set. The test stage here is not really important but it is still
performed.